Funcionalidades http andan correctamente y estan armadas con try y catch, contemplando errores

+++ server.js
@@ -2,7 +2,9 @@ const Container = require("./container.js");
 const productos = new Container("./productos.json");
 const productosEnBase = require("./productos.json");
 
+
 const express = require("express");
+const Router = require("express").Router;
 
 const app = express();
 
@@ -13,24 +15,53 @@ app.use(express.urlencoded({ extended:true }));
 /*Peticiones Get*/
 
 app.get("/api/productos", ( req , res ) => {
-    res.send({ productos: productosEnBase });
+    try{
+        console.log("Se muestran todos los productos correctamente");
+        res.status(200).send({ productos: productosEnBase });
+    }
+    catch(error){
+        console.log("Error en el get de producto");
+        res.status(500).send({ message: error.message })
+    }
 })
 
 app.get("/api/productos/:id", ( req , res ) => {
-    const { id } = req.params;
-    const producto = productosEnBase.find( obj => obj.id == id );
-    res.send({ producto : { producto }}) 
+    try{
+        if (req.params) {
+            const { id } = req.params;
+            const producto = productosEnBase.find( obj => obj.id == id );
+            if (!producto) {
+                console.log("No se encuentra el producto");
+                return res.status(400).send({ error: "Producto no encontrado"});
+            }
+            return res.status(200).send({ producto : { producto }}) 
+        }
+    }catch ( error ) {
+        console.log("Error en el get del producto");
+        res.status(500).send({ message : error.message })
+    }
+    
 })
 
 
 /*Peticiones Post*/
 app.post("/api/productos", ( req , res ) => {
-    const obj = req.body;
-    async function cargarProducto( producto ) {
-        await productos.save( obj );
+    try{
+        if ( req.body.title || req.body.price || req.body.thumbnail ) {
+        const obj = req.body;
+        async function cargarProducto( producto ) {
+            await productos.save( obj );
+        }
+        const nuevoProducto = cargarProducto( obj );
+        res.status(200).send({ nuevoProducto: obj });
     }
-    const nuevoProducto = cargarProducto( obj );
-    res.status(200).send({ nuevoProducto: obj });
+    res.status(200).send({ message:"No hay informacion de usuarios en los params" })
+    }catch ( error ) {
+        console.log("Error en el get del producto");
+        res.status(500).send({ message : error.message })
+    }
+    
+    
 })
 
 /*Peticiones PUT*/
@@ -51,13 +82,23 @@ app.post("/api/productos", ( req , res ) => {
 /*Peticiones DELETE*/
 
 app.delete("/api/productos/:id" , ( req , res ) => {
-    const { id } = req.params;
-    const producto = productosEnBase.find( obj => obj.id == id );
-    async function borrarDeBase(id){
-        await productos.deleteById(id)
+    try {
+        if (req.params) {
+            const { id } = req.params;
+            const producto = productosEnBase.find( obj => obj.id == id );
+            if (!producto) {
+                res.status(404).send({ message: "Ese producto no fue encontrado en la base"})
+            }
+            async function borrarDeBase(id){
+            await productos.deleteById(id)
+            }
+            const productoBorrado = borrarDeBase(id);
+            res.status(200).send({ productoBorrado: producto})
+        }
+    }catch ( error ) {
+            console.log("Error en el get del producto");
+            res.status(500).send({ message : error.message })
     }
-    const productoBorrado = borrarDeBase(id);
-    res.send({ productoBorrado: producto})
 })
